requirements.txt
Django==4.2.2
django-environ==0.10.0
django-filter==23.2
django-modeltranslation==0.18.10
django-rest-auth==0.9.5
djangorestframework==3.14.0
Pillow==9.5.0
redis==4.5.5
celery==5.3.0
requests==2.31.0
drf-yasg==1.21.6
django-cors-headers==4.1.0
coreapi==2.3.3
django-import-export==3.2.0
pyTelegramBotAPI==4.12.0
pandas==2.0.3
pytest-django==4.5.2
model_bakery==1.12.0
